+++ src/Symfony/Bridge/Propel1/Form/DataTransformer/ModelToIdTransformer.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bridge\Propel\Form\DataTransformer;
+namespace Symfony\Bridge\Propel1\Form\DataTransformer;
 
-use Symfony\Bridge\Propel\Form\ChoiceList\ModelChoiceList;
+use Symfony\Bridge\Propel1\Form\ChoiceList\ModelChoiceList;
 use Symfony\Component\Form\DataTransformerInterface;
 use Symfony\Component\Form\Exception\UnexpectedTypeException;
 use Symfony\Component\Form\Exception\TransformationFailedException;
